Move from VS2012 to VS2013
Upgraded to VS2013 and needed to fix project files. Removed namespace
for useless math functions.

+++ UselessNumberMachine/TestBed/TestUselessMath.cpp
@@ -2,7 +2,7 @@
 
 void TestUselessMath::testUnit()
 {
-   int test1;
+   double test1;
    char ch;
    test1 = addition(NumbersList());
 
@@ -13,4 +13,3 @@ void TestUselessMath::testUnit()
    cout << "Press any key to do something else\n";
    cin >> ch;
 }
-
